Revert "Merge branch 'main' into ikotome/front/add-create-page"
This reverts commit 5601f9b0a5de09ab32a25d190d6154acc4e5602a, reversing
changes made to 22f4f58236577d51ee78ebc5ac289aa377d99ca5.
