+++ packages/akane-next/src/app/(main)/(map)/(bottomSheet)/(pageRoot)/story/[story_id]/challenge/page.tsx
