+++ packages/akane-next/src/app/(main)/(map)/(bottomSheet)/layout.tsx
@@ -31,8 +31,9 @@ export default function BottomSheetLayout({
                             position: "fixed",
                             display: "flex",
                             flexDirection: "column",
-                            height: "90%",
+                            height: "calc(100% - 6rem)",
                             width: "100%",
+                            margin: "6rem 6rem 0 0",
                             bottom: 0,
                             left: 0,
                             right: 0,
