+++ packages/akane-next/src/app/(main)/layout.tsx
@@ -1,7 +1,7 @@
 import { BottomSheetProvider } from "@/context/BottomSheetContext";
-import { HeaderQuestionProvider } from "@/context/HeaderQuestionContext";
 import { LocationProvider } from "@/context/LocationContext";
 import { MapProvider } from "@/context/MapContext";
+import { Box } from "@chakra-ui/react";
 
 export default function MainLayout({
     children,
@@ -11,11 +11,16 @@ export default function MainLayout({
     return (
         <MapProvider>
             <LocationProvider>
-                <HeaderQuestionProvider>
-                    <BottomSheetProvider>
+                <BottomSheetProvider>
+                    <Box
+                        pos="relative"
+                        h="100dvh"
+                        w="100vw"
+                        overflow="hidden"
+                    >
                         {children}
-                    </BottomSheetProvider>
-                </HeaderQuestionProvider>
+                    </Box>
+                </BottomSheetProvider>
             </LocationProvider>
         </MapProvider>
     );
