--- packages/akane-next/src/components/mypage/TabTrigger.tsx
@@ -1,17 +1,0 @@
-import { Tabs } from "@chakra-ui/react";
-
-export default function TabTrigger({
-	content,
-	value,
-}: { content: string; value: string }) {
-	return (
-		<Tabs.Trigger
-			value={value}
-			color="#49454F"
-			display="flex"
-			justifyContent="center"
-		>
-			{content}
-		</Tabs.Trigger>
-	);
-}
