+++ packages/akane-next/src/components/story/StoryCard.tsx
@@ -3,7 +3,6 @@ import Link from "next/link";
 import SquareImage from "@/components/story/squareImage";
 import styles from "./StoryCard.module.css";
 import { getCardBorderColor, getStatusMessage } from "@/helper/statusHandler";
-import type { StaticImageData } from "next/image";
 
 /*
 background: カードの背景色：string(Chakra-ui)
@@ -15,77 +14,69 @@ title: カードのタイトル：string
 size: カードのwidthとheight:string(Chakra-ui)
 separatorColor: タイトルと状態を分つホライズンラインの色:string(Chakraui)
 status: カード(story）の進行状態の文字列: string
-	in_progress: "進行中"
-	started: "開始済み"
-	cleared: "完了"
-	public: "投稿済み"
-	url_only: "URLのみ"
-	private: "編集途中"
+  in_progress: "進行中"
+  started: "開始済み"
+  cleared: "完了"
+  public: "投稿済み"
+  url_only: "URLのみ"
+  private: "編集途中"
 */
 export function StoryCard({
-	background = "#f4f4f5",
-	cardBorderWidth = "thick",
-	cardRadius = "2rem",
-	imageURL = "https://images.unsplash.com/photo-1555041469-a586c61ea9bc?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1770&q=80",
-	imageAlt = "this is image for the question.",
-	imageSize = 50,
-	link = "/",
-	title = "タイトル",
-	size = "150px",
-	status = "in_progress",
+  background = "#f4f4f5",
+  cardBorderWidth = "thick",
+  cardRadius = "2rem",
+  imageURL = "https://images.unsplash.com/photo-1555041469-a586c61ea9bc?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1770&q=80",
+  imageAlt = "this is image for the question.",
+  imageSize = 50,
+  link = "/",
+  title = "タイトル",
+  size = "150px",
+  separatorColor = "#49454F",
+  status = "in_progress",
 }: {
-	background: string;
-	cardBorderWidth: string;
-	cardRadius: string;
-	imageURL: string | StaticImageData;
-	imageAlt: string;
-	imageSize: number;
-	link: string;
-	title: string;
-	size: string;
-	status: string;
+  background: string;
+  cardBorderWidth: string;
+  cardRadius: string;
+  imageURL: string;
+  imageAlt: string;
+  imageSize: number;
+  link: string;
+  title: string;
+  size: string;
+  separatorColor: string;
+  status: string;
 }) {
-	const cardBorderColor = getCardBorderColor(status);
-	const statusMessage = getStatusMessage(status);
+  const cardBorderColor = getCardBorderColor(status);
+  const statusMessage = getStatusMessage(status);
 
-	return (
-		<Link href={link}>
-			<Card.Root
-				background={background}
-				borderWidth={cardBorderWidth}
-				borderColor={cardBorderColor}
-				borderRadius={cardRadius}
-				width={size}
-				height={size}
-				overflow="hidden"
-				padding="5px"
-			>
-				<div className={styles.imageContainer}>
-					<SquareImage
-						src={imageURL}
-						alt={imageAlt}
-						size={imageSize}
-						radius={"2xl"}
-					/>
-				</div>
-				<Card.Title
-					marginTop="1.5"
-					textAlign={"center"}
-					color="black"
-					fontWeight="normal"
-				>
-					{title}
-				</Card.Title>
-				<Separator size="xs" colorPalette="gray" />
-				<Text
-					marginTop={"1.5"}
-					textAlign={"center"}
-					color="black"
-					fontWeight="bolder"
-				>
-					{statusMessage}
-				</Text>
-			</Card.Root>
-		</Link>
-	);
+  return (
+    <Link href={link}>
+      <Card.Root
+        background={background}
+        borderWidth={cardBorderWidth}
+        borderColor={cardBorderColor}
+        borderRadius={cardRadius}
+        width={size}
+        height={size}
+        overflow="hidden"
+        padding="5px"
+      >
+        <div className={styles.imageContainer}>
+          <SquareImage
+            src={imageURL}
+            alt={imageAlt}
+            size={imageSize}
+            radius={"2xl"}
+          />
+        </div>
+        <Card.Title marginTop="1.5" textAlign={"center"}>
+          {title}
+        </Card.Title>
+        <Separator size="md" color={separatorColor} />
+        <Text marginTop={"1.5"} textAlign={"center"}>
+          {statusMessage}
+        </Text>
+      </Card.Root>
+    </Link>
+  );
 }
