+++ packages/akane-next/src/constants/status.ts
@@ -1,26 +1,17 @@
 export const CardBorderColor: { [key: string]: string } = {
-	in_progress: "#1591EA",
-	started: "#1591EA",
-	cleared: "#34C759",
-	public: "#A505A5",
-	url_only: "#FF7518",
-	private: "#FF7518",
+  in_progress: "#0000FF",
+  started: "#0000FF",
+  cleared: "#34C759",
+  public: "#7F00FF",
+  url_only: "#FF7518",
+  private: "#FF7518",
 };
 
 export const StatusMessage: { [key: string]: string } = {
-	in_progress: "進行中",
-	started: "開始済み",
-	cleared: "完了",
-	public: "投稿済み",
-	url_only: "URLのみ",
-	private: "編集途中",
-};
-
-export const StoryStatus: { [key: string]: string } = {
-	in_progress: "in_progress",
-	started: "started",
-	cleared: "cleared",
-	public: "public",
-	url_only: "url_only",
-	private: "private",
+  in_progress: "進行中",
+  started: "開始済み",
+  cleared: "完了",
+  public: "投稿済み",
+  url_only: "URLのみ",
+  private: "編集途中",
 };
